### Important Notice: Reflashing the Firmware
Before proceeding with the update, it is imperative to reflash the device firmware with erase (use the webflasher) to avail these new features. Please ensure that you've created a backup of your existing configuration to avoid any loss of your flash files.

### Features and Updates

#### Increase in Free Flash Space
- The new firmware uses a standard partition table, which increases the available flash space for icons from 256KB to 1.36MB. This provides more room for customization.

#### Enhanced Color Depth
- AWTRIX has rewritten its drawing functions to now support 24-bit color depth (16.7 million colors), as opposed to the previous 16-bit (65,536 colors). This will result in richer and more vibrant visuals. App updates relating to this feature are currently under review.

#### Improved WiFi Reconnectivity
- A custom-built WiFi reconnect function replaces the standard framework's reconnect method. AWTRIX will now attempt to reconnect every 30 seconds, and if it fails 10 times, the system will automatically reboot. This ensures a more aggressive and reliable WiFi connectivity.

#### Better GIF Handling
- The new GIF decoder is significantly better at handling compressed GIFs, ensuring smoother animations and graphics.

#### Customizable Calendar Appearance
- Users now have the ability to change the calendar body color using the new `CBCOL` setting. The header color can also be changed using `CHCOL`.

#### UI Improvements
- The system now better handles the background color, progress bars, and the free pixel space between icons and text, offering a cleaner UI.

#### Configuration File Renaming
- The `config.json` file has been renamed to `DoNotTouch.json` to emphasize its importance and discourage accidental modifications.

#### New Lifetime Mode
- Two modes have been added:
  - `(0)` will remove the app from the loop, functioning as before.
  - `(1)` will display a red border around the app to indicate that no updates have been received after a specified lifetime.

#### Native App Name Changes
- The internal names for the native Apps have been changed. This is particularly useful for those using the `/switch` API command.

#### Removal of AWTRIX ID Icon Downloader
- The AWTRIX ID Icon downloader has been removed from the web interface. Users should now use the mobile app for downloading community icons.

#### Display WiFi Network
- The web interface will now display the connected WiFi SSID instead of the IP address, providing more relevant information at a glance.

+++ src/MatrixDisplayUi.h
@@ -136,7 +136,7 @@ private:
 
 public:
   MatrixDisplayUi(FastLED_NeoMatrix *matrix);
-    uint16_t  fadeColor(uint16_t  color, uint32_t interval);
+  uint32_t fadeColor(uint32_t  color, uint32_t interval);
   uint8_t AppCount = 0;
   /**
    * Initialise the display
@@ -177,17 +177,17 @@ public:
    */
   void setTimePerTransition(uint16_t time);
 
-  void setIndicator1Color(uint16_t color);
+  void setIndicator1Color(uint32_t color);
   void setIndicator1State(bool state);
   void setIndicator1Blink(int Blink);
   void setIndicator1Fade(int fade);
 
-  void setIndicator2Color(uint16_t color);
+  void setIndicator2Color(uint32_t color);
   void setIndicator2State(bool state);
   void setIndicator2Blink(int Blink);
   void setIndicator2Fade(int fade);
 
-  void setIndicator3Color(uint16_t color);
+  void setIndicator3Color(uint32_t color);
   void setIndicator3State(bool state);
   void setIndicator3Blink(int Blink);
   void setIndicator3Fade(int fade);
@@ -233,9 +233,9 @@ public:
 
   int8_t update();
 
-  uint16_t indicator1Color = 63488;
-  uint16_t indicator2Color = 31;
-  uint16_t indicator3Color = 65504;
+  uint32_t indicator1Color = 63488;
+  uint32_t indicator2Color = 31;
+  uint32_t indicator3Color = 65504;
 
   bool indicator1State = false;
   bool indicator2State = false;
